+++ src/features/perps/components/PerpMarketRow.tsx
@@ -3,12 +3,12 @@ import React, { useMemo } from 'react';
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { LiveTokenText } from '@/components/live-token-text/LiveTokenText';
 import { Box, Text } from '@/design-system';
-import { usePriceChangeColors } from '@/design-system/color/usePriceChangeColors';
 import { HyperliquidTokenIcon } from '@/features/perps/components/HyperliquidTokenIcon';
 import { LeverageBadge } from '@/features/perps/components/LeverageBadge';
 import { type PerpMarket } from '@/features/perps/types';
 import { formatPriceChange, getHyperliquidTokenId } from '@/features/perps/utils';
 import { formatPerpAssetPrice } from '@/features/perps/utils/formatPerpsAssetPrice';
+import { usePriceChangeColors } from '@/framework/ui/price/usePriceChangeColors';
 import { formatNumber } from '@/helpers/strings';
 import * as i18n from '@/languages';
 import { type TokenData } from '@/state/liveTokens/liveTokensStore';
